+++ src/crypto/key_exchange.h
@@ -3,8 +3,8 @@
 #pragma once
 
 #include "ccf/crypto/key_pair.h"
+#include "ccf/crypto/openssl/openssl_wrappers.h"
 #include "ccf/ds/logger.h"
-#include "crypto/openssl/openssl_wrappers.h"
 #include "crypto/openssl/public_key.h"
 
 #include <iostream>