+++ src/crypto/openssl/key_pair.h
@@ -3,9 +3,9 @@
 #pragma once
 
 #include "ccf/crypto/key_pair.h"
+#include "ccf/crypto/openssl/openssl_wrappers.h"
 #include "ccf/crypto/public_key.h"
 #include "crypto/openssl/public_key.h"
-#include "openssl_wrappers.h"
 
 #include <optional>
 #include <stdexcept>